refactor(HitCanvas): Extract `rgbColorGenerator()` and `getPixelColor()` utils
